requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
PyYAML>=6.0.1
click>=8.1.7
lxml>=4.9.0
urllib3>=2.0.0
python-magic>=0.4.27
pillow>=10.0.0
networkx>=3.2.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

